project/app/requirements.txt
asgiref==3.5.2
attrs==22.2.0
Django==3.2.13
django-split-settings==1.2.0
djangorestframework==3.14.0
django_debug_toolbar==3.8.1
drf-spectacular==0.25.1
flake8==6.0.0
inflection==0.5.1
jsonschema==4.17.3
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.2
pyflakes==3.0.1
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
gunicorn==20.0.4
elasticsearch==7.7.0

